+++ electron/constants/index.js
@@ -1,0 +1,54 @@
+const os = require('os');
+const path = require('path');
+
+const PORT_RANGE = { startPort: 39152, endPort: 65535 };
+const ERROR_ADDRESS_IN_USE = 'EADDRINUSE';
+
+// OS specific constants
+const isWindows = process.platform === 'win32';
+const isMac = process.platform === 'darwin';
+const isLinux = process.platform === 'linux';
+
+// Environment specific constants
+const isProd = process.env.NODE_ENV === 'production';
+const isDev = !isProd;
+
+// Paths
+const dotOperateDirectory = isProd
+  ? path.join(os.homedir(), '.operate')
+  : '.operate';
+
+const paths = {
+  dotOperateDirectory,
+  servicesDir: path.join(dotOperateDirectory, 'services'),
+  venvDir: path.join(dotOperateDirectory, 'venv'),
+  tempDir: path.join(dotOperateDirectory, 'temp'),
+  versionFile: path.join(dotOperateDirectory, 'version.txt'),
+  cliLogFile: path.join(dotOperateDirectory, 'cli.log'),
+  electronLogFile: path.join(dotOperateDirectory, 'electron.log'),
+  nextLogFile: path.join(dotOperateDirectory, 'next.log'),
+  osPearlTempDir: path.join(os.tmpdir(), 'pearl'),
+};
+
+// Publish options
+const publishOptions = {
+  provider: 'github',
+  owner: 'valory-xyz',
+  repo: 'olas-operate-app',
+  releaseType: 'draft',
+  token: process.env.GH_TOKEN,
+  private: false,
+  publishAutoUpdate: false,
+};
+
+module.exports = {
+  PORT_RANGE,
+  ERROR_ADDRESS_IN_USE,
+  isWindows,
+  isMac,
+  isLinux,
+  isProd,
+  isDev,
+  publishOptions,
+  paths,
+};
