--- electron/constants/publishOptions.js
@@ -1,11 +1,0 @@
-const publishOptions = {
-  provider: 'github',
-  owner: 'valory-xyz',
-  repo: 'olas-operate-app',
-  releaseType: 'draft',
-  token: process.env.GH_TOKEN,
-  private: false,
-  publishAutoUpdate: false,
-};
-
-module.exports = { publishOptions };
